+++ kernels/gemm/gemm_fp8fp4_gfx1250.py
@@ -19,6 +19,7 @@ from flydsl.expr.typing import T
 from flydsl.runtime.device import get_rocm_arch
 from flydsl.utils.smem_allocator import SmemAllocator, SmemPtr, check_smem_capacity
 from kernels.common import buffer_ops
+from kernels.common.gfx1250_cluster import compute_mcast_masks
 from kernels.common.mma.pipeline_utils_gfx1250 import make_tail_plan, tdm_epilogue_fence_threshold_bytes
 from kernels.common.utils import align_up as _align_up
 from kernels.gemm.gemm_common_gfx1250 import (
@@ -511,7 +512,7 @@ def compile_fp8fp4_gemm(
 
         if const_expr(use_cluster):
             local_x, local_y = cluster.compute_cluster_position()
-            a_mcast_mask, b_mcast_mask = cluster.compute_mcast_masks(local_x, local_y, cluster_m, cluster_n)
+            a_mcast_mask, b_mcast_mask = compute_mcast_masks(local_x, local_y, cluster_m, cluster_n)
         else:
             a_mcast_mask = 0
             b_mcast_mask = 0
